fastapi_base/requirements.txt
better_exceptions
httpx~=0.23.3
PyExecJS
APScheduler~=3.10.0
loguru~=0.6.0
fastapi~=0.92.0
pycryptodome
uvicorn~=0.20.0
aiomysql~=0.1.1
asyncio~=3.4.3
pydantic~=1.10.5
starlette~=0.25.0
PyJWT~=2.7.0
PyYAML~=6.0
PyMySQL~=1.0.2
DBUtils~=3.0.2
redis~=2.10.6
aioredis~=2.0.1